requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
pandas>=2.0.0
chardet>=5.2.0
tqdm>=4.66.0
chromadb>=0.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
openai>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
email-validator>=2.0.0
streamlit>=1.28.0
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0
langchain-google-genai>=0.0.6
langchain-community>=0.4.0
langgraph>=1.0.0

